pyproject.toml
[tool.poetry]
name = "PyPwnCollege"
packages = [
    {include = "pwncollege"}
]
version = "0.1.0"
description = "A wrapper for the pwn.college API."
authors = ["Zeeshan Ayub <mayub9799@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]
Sphinx = "^4.5.0"
pytest = "^7.1.2"
python-dotenv = "^0.21.0"
pytest-cov = "^3.0.0"
pytest-xdist = "^2.5.0"
sphinx-rtd-theme = "^1.0.0"
mypy = "^0.950"
types-requests = "^2.27.25"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
